[FIX] stock_picking_mgmt_weight: ensure no cancelled moves when SO confirmation
For proper cancel and pending quantities calculation, a SO should only have
cancelled stock moves that belong to cancel pending quantities operation.
With this fix, when a SO is confirmed, every cancelled move is erased to
ensure no previous cancelled moves exist, taht would break pending and cancel
quantites calculation.