[IMP] stock_valuation: enable history average stock quantity edition
This improvement, as expected, fires general stock recalculation for later PHAPs.
The differente quantity between current and desired is assigned at date
to the default stock location for the selected warehouse.
This is because this view is for a selected warehouse, not location, and at this
point it's not possible selecting any location within the warehouse.