[IMP] stock_picking_mgmt_weight: ticket creation - use net date for numbering
With this improvement, in report ticket proper date (tare for incoming weights
and gross date for outgoing) is printed for general ticket print date.

Also, this date is used for ticket numbering, when date range is used in
the correspondent stock picking sequence. This allows to create weight tickets
in the past (e.g. last year) with the desired number according this date.