[ADD] stock_move_line_lot_qty_total
Enables configuring desired products tracked by lot to be delivered using
current available stock.

This is achieved enabling for a selected, lot tracked product, the
Force available quantity in deliveries check. Then, for an outgoing picking
line in a pending state, its possible to manually select destination lots and,
then, clicking in fill available stock for already filled lots.

When a picking is validated a current stock for lots is checked again,
preventing mark as done if a lot is filled with a done quantity different than
available one.