[FIX] stock_valuation: proper independent outgoing pickings management
Before this fix, standalone outgoing pickings, that are not linked
to sales, or purchase returns, were not properly managed. SVL records
were treated as "accumulated", with zero price, then average price was
wrongly incremented.
This issue, also raised an error for some limit cases, due to a division
by zero (when the outgoing move set to 0 warehouse stock and create a new
PHAP day record as well).

This fix makes these outputs work like e.g. a Sales output, that is a
simple case (valuated as average price of the output day)